Bug on createroom , Cant do anything with DEFAULT

+++ backend/src/whiteboard.controller.ts
@@ -8,44 +8,44 @@ export class WhiteboardController {
   constructor(private readonly whiteboardService: WhiteboardService) { }
 
   @Post('createcatagories')
-  async createCatagories(@Body() data:{name:string}) {
+  async createCatagories(@Body() data: { name: string }) {
     const creating = this.whiteboardService.createCatagories(data)
     return creating
   }
 
   @Post('createroom')
-  async createRoom(@Body() data:{member:string , memberName:string ,catagorie:string , roomname:string}) {
+  async createRoom(@Body() data: { member: string, memberName: string, catagorie: string, roomname: string }) {
     //console.log('creating');
     const creating = this.whiteboardService.createRoom(data)
     return creating
   }
 
   @Put('editcatagories')
-  async editCatagories(@Body() data:{oldname:string , newname:string}) {
+  async editCatagories(@Body() data: { oldname: string, newname: string }) {
     const editing = this.whiteboardService.editCatagories(data)
     return editing
   }
 
   @Put('changecatagories')
-  async changeCatagories(@Body() data:{room:string , from:string , to:string}) {
-    const changing = this.whiteboardService.changeCatagories(data)
-    return changing
+  async changeCatagories(@Body() data: { room: string, from: string, to: string }) {
+    if (data.from != 'DEFAULT') {
+      const changing = this.whiteboardService.changeCatagories(data)
+      return changing
+    }
   }
 
-  @Post('deleteroom') 
-  async deleteRoom(@Body() data:{catagorie:string , room:string}) {
-    console.log('delete',data);
-    
+  @Post('deleteroom')
+  async deleteRoom(@Body() data: { catagorie: string, room: string }) {
     const deleting = this.whiteboardService.deleteRoom(data)
     return deleting
   }
 
   @Post('deleteCatagories')
-  async deleteCatagories(@Body() data:{catagories:string}) {
-    console.log("->",data);
-    
-    const deleting = this.whiteboardService.deleteCatagories(data)
-    return deleting
+  async deleteCatagories(@Body() data: { catagories: string }) {
+    if (data.catagories != "DEFAULT") {
+      const deleting = this.whiteboardService.deleteCatagories(data)
+      return deleting
+    }
   }
   // @Post('addmember')
   // async addMember(@Body() data:{room:string , member:string}) {
